address version display review feedback
