+++ web/vite.user.config.ts
@@ -1,26 +1,9 @@
-import { execFileSync } from "node:child_process";
 import { resolve } from "node:path";
 import { readFileSync } from "node:fs";
 import { defineConfig } from "vitest/config";
 import react from "@vitejs/plugin-react";
 import type { Plugin } from "vite";
-
-function denseMemVersion() {
-  const configuredVersion = process.env.DENSE_MEM_VERSION ?? process.env.VITE_DENSE_MEM_VERSION ?? process.env.IMAGE_VERSION;
-  if (configuredVersion?.trim()) {
-    return configuredVersion.trim();
-  }
-
-  try {
-    const version = execFileSync("git", ["describe", "--tags", "--always", "--dirty"], {
-      cwd: resolve(__dirname, ".."),
-      stdio: ["ignore", "pipe", "ignore"],
-    }).toString().trim();
-    return version || "dev";
-  } catch {
-    return "dev";
-  }
-}
+import { denseMemVersion } from "./vite.version";
 
 function serveUserPortalAtUi(): Plugin {
   return {
